websockets/requirements.txt
websockets==11.0.3
PyJWT == 2.6.0
SQLAlchemy == 2.0.17
psycopg2-binary == 2.9.6
pydantic == 1.10.5
python-dotenv == 0.21.1
